project_manager/requirements.txt
asgiref==3.5.0
#bleach==6.0.0
blinker==1.4
#certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.25
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.0 ##################################3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
dnspython==2.2.0
docker==6.1.3
eventlet==0.33.0
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
jwcrypto==1.0
kaggle==1.6.14
mongoengine==0.24.0
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0 ###################################3
opencv-python==4.7.0.72
packaging==24.0
Pillow==9.5.0
#pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pymongo==4.0
PyMySQL==1.0.2
#python-dateutil==2.9.0.post0
#python-slugify==8.0.4
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
#text-unidecode==1.3
#tqdm==4.66.4
typing_extensions==4.1.1
urllib3==1.26.8
#webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.0.3
wrapt==1.14.0
yarl==1.7.2
pandas===1.3.5
openpyxl===3.1.3
uWSGI==2.0.24
wheel==0.42.0